requirements.txt
biopython==1.86
dotenv==0.9.9
grpcio==1.76.0
grpcio-tools==1.76.0
numpy==2.3.4
protobuf==6.33.1
python-dotenv==1.2.1
setuptools==80.9.0
typing_extensions==4.15.0
